Applied patch from Hermann for correct Envelope behaviour. Changed Master volume and fine detune to parcombi.     Detune required a change in Interchange.cpp as indirectTransfers     assumed all values were integers. Changed Default and Current theme boxes to unresponsive buttons.     This was so long names are clipped and don't spill over Updated Envelope Limits. Updated LFO Limits. Updated Filter Limits. Applied corrections to User Guide. Set version as 2.3.0 rc6

+++ src/Synth/Envelope.h
@@ -45,8 +45,8 @@ class Envelope
     private:
         EnvelopeParams *_envpars;
         Presets::PresetsUpdate envUpdate;
-        int envpoints;
-        int envsustain;   // 0 means disabled
+        size_t envpoints;
+        size_t envsustain;   // 0 means disabled
         float envval[MAX_ENVELOPE_POINTS]; // [0.0 .. 1.0]
         float envstretch;
         int linearenvelope;
@@ -55,7 +55,7 @@ class Envelope
         int forcedrelase;
         char keyreleased;    // if the key was released
         char envfinish;
-        int currentpoint;    // current envelope point (starts from 1)
+        size_t currentpoint; // current envelope point (starts from 1)
         float t;             // the time from the last point
         float envoutval;     // used to do the forced release
